requirements.txt
fastapi
uvicorn
jinja2
aiofiles
python-multipart
passlib[bcrypt]
sqlmodel
python-dotenv
